update: Now, non-logged-in users can shorten their URLs, but can't use exclusive features. add features: Provides usage statistics for each shortened URL, such as such as number of clicks and creation date.

+++ index.php
@@ -16,18 +16,12 @@ function generateShortURL()
     return $shortURL;
 }
 
-// Check if user is not logged in yet, redirect to login page
-if (!isset($_SESSION['user_id'])) {
-    header('Location: auth/login.php');
-    exit();
-}
-
 // Handle the URL shortening process
 $error = '';
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $originalURL = filter_input(INPUT_POST, 'url', FILTER_SANITIZE_URL);
     $customURL = filter_input(INPUT_POST, 'custom_url', FILTER_SANITIZE_STRING);
-    $userId = $_SESSION['user_id'];
+    $userId = isset($_SESSION['user_id']) ? $_SESSION['user_id'] : null;
 
     // Filtering inserted URL
     if (!filter_var($originalURL, FILTER_VALIDATE_URL)) {
@@ -78,68 +72,53 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     }
 }
 
-// Retrieve the shortened URL for the current user
-$conn = connectToDatabase();
-
-$stmt = $conn->prepare("SELECT id, short_url, original_url, created_at FROM url_mappings WHERE user_id = ?");
-
-if (!$stmt) {
-    die("Preparation of prepared statement failed: " . $conn->error);
-}
-
-$stmt->bind_param("i", $_SESSION['user_id']);
-
-if ($stmt->execute()) {
-    $result = $stmt->get_result();
-    $archivedURLs = $result->fetch_all(MYSQLI_ASSOC);
-} else {
-    die("Execution of prepared statement failed: " . $stmt->error);
-}
-
-$stmt->close();
-$conn->close();
-
-// Handle URL deletion
-if (isset($_GET['delete']) && isset($_GET['id'])) {
-    $deleteId = $_GET['id'];
+// Retrieve the shortened URL for the current user (if logged in)
+$archivedURLs = array();
+if (isset($_SESSION['user_id'])) {
     $conn = connectToDatabase();
-    $stmt = $conn->prepare("DELETE FROM url_mappings WHERE id = ? AND user_id = ?");
+
+    $stmt = $conn->prepare("SELECT id, short_url, original_url, views, created_at FROM url_mappings WHERE user_id = ?");
+
     if (!$stmt) {
-        $error = "Preparation of prepared statement failed: " . $conn->error;
-    } else {
-        $stmt->bind_param("ii", $deleteId, $_SESSION['user_id']);
-        if ($stmt->execute()) {
-            header("Location: index.php");
-            exit();
-        } else {
-            $error = "Execution of prepared statement failed: " . $stmt->error;
-        }
-        $stmt->close();
+        die("Preparation of prepared statement failed: " . $conn->error);
     }
+
+    $stmt->bind_param("i", $_SESSION['user_id']);
+
+    if ($stmt->execute()) {
+        $result = $stmt->get_result();
+        $archivedURLs = $result->fetch_all(MYSQLI_ASSOC);
+    } else {
+        die("Execution of prepared statement failed: " . $stmt->error);
+    }
+
+    $stmt->close();
     $conn->close();
 }
+
 ?>
 
 <!DOCTYPE html>
 <html>
 <head>
     <title>URL Shortener</title>
-    <!-- Used to control the appearance of web pages to fit the screen width of the user's device -->
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <!-- Add Bootstrap CSS link -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
-    <!-- Specifies the character encoding of the web page (usually using UTF-8). -->
-    <meta charset="utf-8">
-    <!-- Provides a brief description of the content of a web page for search engine purposes -->
-    <meta name="description" content="URL Shortener is a tool to shorten long URLs and make them more manageable.">
-    <!-- Determines keywords related to web pages for search engine purposes -->
-    <meta name="keywords" content="URL shortener, short URLs, link shortener, web tools">
+        <!-- Used to control the appearance of web pages to fit the screen width of the user's device -->
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <!-- Add Bootstrap CSS link -->
+        <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+        <!-- Specifies the character encoding of the web page (usually using UTF-8). -->
+        <meta charset="utf-8">
+        <!-- Provides a brief description of the content of a web page for search engine purposes -->
+        <meta name="description" content="URL Shortener is a tool to shorten long URLs and make them more manageable.">
+        <!-- Determines keywords related to web pages for search engine purposes -->
+        <meta name="keywords" content="URL shortener, short URLs, link shortener, web tools">
 </head>
 <body>
     <div class="container">
         <h1 class="mb-4">URL Shortener</h1>
-        <p class="mb-4">Welcome, <?php echo htmlspecialchars($_SESSION['username']); ?>! <a href="auth/logout.php">Logout</a></p>
-
+        <?php if (isset($_SESSION['user_id'])) { ?>
+            <p class="mb-4">Welcome, <?php echo htmlspecialchars($_SESSION['username']); ?>! <a href="auth/logout.php">Logout</a></p>
+        <?php } ?>
         <div class="card shorten-form">
             <div class="card-header">
                 <h2>Shorten a URL</h2>
@@ -150,16 +129,27 @@ if (isset($_GET['delete']) && isset($_GET['id'])) {
                         <label for="url" class="form-label">URL:</label>
                         <input type="text" class="form-control" name="url" id="url" placeholder="Enter URL" required>
                     </div>
-                    <div class="mb-3">
-                        <label for="custom_url" class="form-label">Custom URL (optional):</label>
-                        <input type="text" class="form-control" name="custom_url" id="custom_url" placeholder="Enter custom URL">
-                    </div>
+                    <?php if (isset($_SESSION['user_id'])) { ?>
+                        <div class="mb-3">
+                            <label for="custom_url" class="form-label">Custom URL (optional):</label>
+                            <input type="text" class="form-control" name="custom_url" id="custom_url" placeholder="Enter custom URL">
+                        </div>
+                    <?php } ?>
                     <button type="submit" class="btn btn-primary">Shorten</button>
                 </form>
                 <?php if (!empty($error)) { ?>
                     <div class="alert alert-danger mt-4">
                         <?php echo htmlspecialchars($error); ?>
                     </div>
+                <?php } ?>
+                <?php if (!isset($_SESSION['user_id'])) { ?>
+                    <p class="mt-4">[!] You can <a href="auth/login.php">login</a> to access exclusive features such as:</p>
+                    <ul>
+                        <li>View the short URLs that you have created</li>
+                        <li>Manage short URLs that you have created</li>
+                        <li>Able to customize short URLs as you want</li>
+                        <li>And more...</li>
+                    </ul>
                 <?php } ?>
             </div>
         </div>
@@ -171,38 +161,43 @@ if (isset($_GET['delete']) && isset($_GET['id'])) {
             </div>
         <?php } ?>
 
-        <div class="archived-urls">
-            <h2 class="mt-5">Archived URLs</h2>
-            <?php if (count($archivedURLs) > 0) { ?>
-                <table class="table mt-3">
-                    <thead>
-                        <tr>
-                            <th>Short URL</th>
-                            <th>Original URL</th>
-                            <th>Created At</th>
-                            <th>Action</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <?php foreach ($archivedURLs as $archivedURL) { ?>
+        <?php if (isset($_SESSION['user_id'])) { ?>
+            <div class="archived-urls">
+                <h2 class="mt-5">Archived URLs</h2>
+                <?php if (count($archivedURLs) > 0) { ?>
+                    <table class="table mt-3">
+                        <thead>
                             <tr>
-                                <td><a href="<?php echo htmlspecialchars($archivedURL['short_url']); ?>"><?php echo htmlspecialchars($archivedURL['short_url']); ?></a></td>
-                                <td><?php echo htmlspecialchars($archivedURL['original_url']); ?></td>
-                                <td><?php echo htmlspecialchars($archivedURL['created_at']); ?></td>
-                                <td>
-                                    <a href="index.php?delete=true&id=<?php echo htmlspecialchars($archivedURL['id']); ?>" class="btn btn-danger" onclick="return confirm('Are you sure you want to delete this URL?')">Delete</a>
-                                </td>
+                                <th>Short URL</th>
+                                <th>Original URL</th>
+                                <th>Views</th>
+                                <th>Created At</th>
+                                <th>Action</th>
                             </tr>
-                        <?php } ?>
-                    </tbody>
-                </table>
-            <?php } else { ?>
-                <p>No URLs found.</p>
-            <?php } ?>
-        </div>
+                        </thead>
+                        <tbody>
+                            <?php foreach ($archivedURLs as $archivedURL) { ?>
+                                <tr>
+                                    <td><a href="<?php echo htmlspecialchars($archivedURL['short_url']); ?>"><?php echo htmlspecialchars($archivedURL['short_url']); ?></a></td>
+                                    <td><?php echo htmlspecialchars($archivedURL['original_url']); ?></td>
+                                    <td><?php echo htmlspecialchars($archivedURL['views']); ?></td>
+                                    <td><?php echo htmlspecialchars($archivedURL['created_at']); ?></td>
+                                    <td>
+                                        <a href="index.php?delete=true&id=<?php echo htmlspecialchars($archivedURL['id']); ?>" class="btn btn-danger" onclick="return confirm('Are you sure you want to delete this URL?')">Delete</a>
+                                    </td>
+                                </tr>
+                            <?php } ?>
+                        </tbody>
+                    </table>
+                <?php } else { ?>
+                    <p>No URLs found.</p>
+                <?php } ?>
+            </div>
+        <?php } ?>
     </div>
     <!-- Add Bootstrap JS scripts (jQuery and Bootstrap) -->
     <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.2/dist/js/bootstrap.min.js"></script>
 </body>
+</html>
 </html>
